refactor: unify participant form
