+++ src/app/[id]/partners/components/Modals/ModalParticipants/index.tsx
@@ -2,10 +2,9 @@
 'use client';
 
 import React from 'react';
-import {ArrowLeftIcon} from '@heroicons/react/24/outline';
-import FormPF from './FormPF';
-import FormPJ from './FormPJ';
-import {MemberNode} from '@/types/Members';
+import { ArrowLeftIcon } from '@heroicons/react/24/outline';
+import FormParticipant from './FormParticipant';
+import { MemberNode } from '@/types/Members';
 
 export type ModalPaticipantsProps = {
   open: boolean;
@@ -19,48 +18,17 @@ export type ModalPaticipantsProps = {
   parentBusinessId?: string;
 };
 
-function mapMemberToPF(data?: Partial<MemberNode>) {
-  return {
-    id: data?.id ?? '',                         // importante p/ PUT
-    document: data?.details?.document ?? '',
-    name: data?.details?.name ?? '',
-    percentage: data?.participation_percentage ?? '',
-    representative: data?.type === 'LEGAL_REPRESENTATIVE',
-  };
-}
-
-function mapMemberToPJ(data?: Partial<MemberNode>) {
-  return {
-    id: data?.id ?? '',                         // importante p/ PUT
-    corporate_name: data?.details?.name ?? '',
-    cnpj: data?.details?.document ?? '',
-    percentage: data?.participation_percentage ?? '',
-  };
-}
-
 const ModalPaticipants: React.FC<ModalPaticipantsProps> = ({
-                                                             open,
-                                                             mode,
-                                                             initialData,
-                                                             lockType,
-                                                             onClose,
-                                                             onSaved,
-                                                             clientId,
-                                                             targetLevel,
-                                                             parentBusinessId,
-                                                           }) => {
-  const activeType: 'PERSON' | 'BUSINESS' = (initialData?.member_type as any) ?? 'PERSON';
-  const [activeTab, setActiveTab] = React.useState<'PERSON' | 'BUSINESS'>(activeType);
-
-  const onTabClick = (tab: 'PERSON' | 'BUSINESS') => (e: React.MouseEvent<HTMLAnchorElement>) => {
-    e.preventDefault();
-    if (lockType || mode === 'edit') return;
-    setActiveTab(tab);
-  };
-
-  const isActive = (tab: 'PERSON' | 'BUSINESS') => activeTab === tab;
-
-  // O pai só precisa fechar/atualizar quando o filho terminar:
+  open,
+  mode,
+  initialData,
+  lockType,
+  onClose,
+  onSaved,
+  clientId,
+  targetLevel,
+  parentBusinessId,
+}) => {
   const handleSavedFromChild = (saved: MemberNode) => {
     onSaved?.(saved);
     onClose();
@@ -68,94 +36,68 @@ const ModalPaticipants: React.FC<ModalPaticipantsProps> = ({
 
   if (!open) return null;
 
+  const initialValues = initialData
+    ? {
+        id: initialData.id,
+        name: initialData.details?.name,
+        document: initialData.details?.document,
+        percentage: initialData.participation_percentage,
+        representative: initialData.type === 'LEGAL_REPRESENTATIVE',
+        member_type: initialData.member_type as 'PERSON' | 'BUSINESS',
+        level: initialData.level,
+        parent_business_id: initialData.parent_business_id ?? null,
+      }
+    : undefined;
+
   return (
-      <div className="w-full h-full fixed top-0 left-0 bg-[#F8F9FB] overflow-auto z-[30]">
-        <div className="max-w-4xl h-full mx-auto">
-          <div className="w-full flex justify-center py-6">
-            <div className="w-2/12">
-              <button
-                  className="p-1.5 inline-flex items-center rounded-full bg-gray/20 hover:text-zinc-900"
-                  onClick={onClose}
-                  type="button"
-              >
-                <ArrowLeftIcon className="h-4 w-4 stroke-neutral" />
-              </button>
-            </div>
-
-            <h2 className="flex-1 text-center font-bold">
-              {mode === 'edit' ? 'Editar Participante' : 'Adicionar Participante'}
-            </h2>
-
-            <div className="w-2/12" />
+    <div className="w-full h-full fixed top-0 left-0 bg-[#F8F9FB] overflow-auto z-[30]">
+      <div className="max-w-4xl h-full mx-auto">
+        <div className="w-full flex justify-center py-6">
+          <div className="w-2/12">
+            <button
+              className="p-1.5 inline-flex items-center rounded-full bg-gray/20 hover:text-zinc-900"
+              onClick={onClose}
+              type="button"
+            >
+              <ArrowLeftIcon className="h-4 w-4 stroke-neutral" />
+            </button>
           </div>
 
-          <div className="w-full flex-1 bg-white border border-gray/10 shadow-lg px-4 rounded-t-2xl">
-            <div className="w-full max-w-xl mx-auto">
-              {mode === 'create' && (
-                  <>
-                    <div className="flex items-center justify-center gap-2 pt-4 pb-6">
-                      <a
-                          href="#"
-                          onClick={onTabClick('PERSON')}
-                          className={`px-3 py-1.5 rounded-full text-sm ${isActive('PERSON') ? 'bg-black text-white' : 'bg-neutral-100 text-neutral-700'}`}
-                      >
-                        Pessoa Física
-                      </a>
-                      <a
-                          href="#"
-                          onClick={onTabClick('BUSINESS')}
-                          className={`px-3 py-1.5 rounded-full text-sm ${isActive('BUSINESS') ? 'bg-black text-white' : 'bg-neutral-100 text-neutral-700'}`}
-                      >
-                        Pessoa Jurídica
-                      </a>
-                    </div>
+          <h2 className="flex-1 text-center font-bold">
+            {mode === 'edit' ? 'Editar Participante' : 'Adicionar Participante'}
+          </h2>
 
-                    <div className={isActive('PERSON') ? '' : 'hidden'}>
-                      <FormPF
-                          clientId={clientId}
-                          level={targetLevel}
-                          parentBusinessId={parentBusinessId}
-                          initialValues={undefined}
-                          onSaved={handleSavedFromChild} // <- o filho chama quando terminar
-                      />
-                    </div>
-                    <div className={isActive('BUSINESS') ? '' : 'hidden'}>
-                      <FormPJ
-                          clientId={clientId}
-                          level={targetLevel}
-                          parentBusinessId={parentBusinessId}
-                          initialValues={undefined}
-                          onSaved={handleSavedFromChild}
-                      />
-                    </div>
-                  </>
-              )}
+          <div className="w-2/12" />
+        </div>
 
-              {mode === 'edit' && activeType === 'PERSON' && (
-                  <FormPF
-                      clientId={clientId}
-                      level={initialData?.level}
-                      parentBusinessId={initialData?.parent_business_id ?? undefined}
-                      initialValues={mapMemberToPF(initialData)}
-                      readOnlyType
-                      onSaved={handleSavedFromChild}
-                  />
-              )}
+        <div className="w-full flex-1 bg-white border border-gray/10 shadow-lg px-4 rounded-t-2xl">
+          <div className="w-full max-w-xl mx-auto">
+            {mode === 'create' && (
+              <FormParticipant
+                clientId={clientId}
+                level={targetLevel}
+                parentBusinessId={parentBusinessId}
+                mode="create"
+                readOnlyType={lockType}
+                onSaved={handleSavedFromChild}
+              />
+            )}
 
-              {mode === 'edit' && activeType === 'BUSINESS' && (
-                  <FormPJ
-                      clientId={clientId}
-                      level={initialData?.level}
-                      parentBusinessId={initialData?.parent_business_id ?? undefined}
-                      initialValues={mapMemberToPJ(initialData)}
-                      readOnlyType
-                      onSaved={handleSavedFromChild}
-                  />
-              )}
-            </div>
+            {mode === 'edit' && (
+              <FormParticipant
+                clientId={clientId}
+                level={initialData?.level}
+                parentBusinessId={initialData?.parent_business_id ?? undefined}
+                mode="edit"
+                readOnlyType
+                initialValues={initialValues}
+                onSaved={handleSavedFromChild}
+              />
+            )}
           </div>
         </div>
       </div>
+    </div>
   );
 };
